refactor(AYC-77): extract shared usage dashboard widgets
- Extract model-distribution-chart widget with shared chart-building logic
- Extract provider-consumption-chart widget with time-series transformation
- Extract user-usage-table widget with pagination and row components
- Extract usage-stats-cards widget with loading skeleton
- Simplify 9 page-level components to thin data-fetching wrappers
- Eliminate all cross-page imports violating FSD architecture
- Fix pre-existing lint errors in LongChatWarning and alert components
- Fix complexity violations in UserUsageTableContent by splitting into smaller units
