+++ ayunis-core-frontend/src/pages/admin-settings/usage-settings/ui/provider-consumption-chart/ProviderConsumption.tsx
@@ -5,11 +5,8 @@ import { useMemo } from 'react';
 import { useProviderUsageChart } from '@/pages/admin-settings/usage-settings/api';
 import { useProviders } from '@/features/models';
 
-// UI
-import { ProviderConsumptionLoading } from './ProviderConsumptionLoading';
-import { ProviderConsumptionError } from './ProviderConsumptionError';
-import { ProviderConsumptionEmpty } from './ProviderConsumptionEmpty';
-import { ProviderConsumptionChart } from './ProviderConsumptionChart';
+// Widgets
+import { ProviderConsumptionWidget } from '@/widgets/provider-consumption-chart';
 
 interface ProviderConsumptionProps {
   startDate?: Date;
@@ -34,7 +31,6 @@ export function ProviderConsumption({
 
   const { providers } = useProviders();
 
-  // Create a map from technical provider name to display name
   const providerDisplayNames = useMemo(() => {
     const map: Record<string, string> = {};
     providers.forEach((p) => {
@@ -43,48 +39,12 @@ export function ProviderConsumption({
     return map;
   }, [providers]);
 
-  const { chartData, chartConfig } = useMemo(() => {
-    const empty = {
-      chartData: [] as Array<Record<string, string | number>>,
-      chartConfig: {} as Record<string, { label: string; color: string }>,
-    };
-    const rows = chartResp?.timeSeries ?? [];
-    if (rows.length === 0) return empty;
-
-    const seriesKeys = Object.keys(rows[0].values ?? {});
-    const chartData = rows.map((r) => ({ date: r.date, ...r.values }));
-
-    const palette = [
-      'var(--chart-1)',
-      'var(--chart-2)',
-      'var(--chart-3)',
-      'var(--chart-4)',
-      'var(--chart-5)',
-    ];
-    const chartConfig: Record<string, { label: string; color: string }> = {};
-    seriesKeys.forEach((key, idx) => {
-      chartConfig[key] = {
-        label: providerDisplayNames[key] || key,
-        color: palette[idx % palette.length],
-      };
-    });
-
-    return { chartData, chartConfig };
-  }, [chartResp?.timeSeries, providerDisplayNames]);
-
-  if (isLoading) {
-    return <ProviderConsumptionLoading />;
-  }
-
-  if (error) {
-    return <ProviderConsumptionError error={error} />;
-  }
-
-  if (!chartData || chartData.length === 0) {
-    return <ProviderConsumptionEmpty />;
-  }
-
   return (
-    <ProviderConsumptionChart chartData={chartData} chartConfig={chartConfig} />
+    <ProviderConsumptionWidget
+      timeSeries={chartResp?.timeSeries}
+      providerDisplayNames={providerDisplayNames}
+      isLoading={isLoading}
+      error={error}
+    />
   );
 }
