+++ ayunis-core-frontend/src/pages/super-admin-settings/org/ui/SuperAdminUsageFilters.tsx
@@ -59,8 +59,8 @@ export function SuperAdminUsageFilters({
     if (!permittedModels?.length) return [];
     const modelMap = new Map<string, string>();
     permittedModels.forEach((m) => {
-      if (!modelMap.has(m.name)) {
-        modelMap.set(m.name, m.displayName);
+      if (!modelMap.has(m.id)) {
+        modelMap.set(m.id, m.displayName);
       }
     });
     return Array.from(modelMap.entries()).map(([value, label]) => ({
