+++ ayunis-core-frontend/src/shared/ui/shadcn/alert.tsx
@@ -1,7 +1,7 @@
-import * as React from 'react'
-import { cva, type VariantProps } from 'class-variance-authority'
+import * as React from 'react';
+import { cva, type VariantProps } from 'class-variance-authority';
 
-import { cn } from '@/shared/lib/shadcn/utils'
+import { cn } from '@/shared/lib/shadcn/utils';
 
 const alertVariants = cva(
   'relative w-full rounded-lg border px-4 py-3 text-sm grid has-[>svg]:grid-cols-[calc(var(--spacing)*4)_1fr] grid-cols-[0_1fr] has-[>svg]:gap-x-3 gap-y-0.5 items-start [&>svg]:size-4 [&>svg]:translate-y-0.5 [&>svg]:text-current',
@@ -18,8 +18,8 @@ const alertVariants = cva(
     defaultVariants: {
       variant: 'default',
     },
-  }
-)
+  },
+);
 
 function Alert({
   className,
@@ -34,7 +34,7 @@ function Alert({
       className={cn(alertVariants({ variant }), className)}
       {...props}
     />
-  )
+  );
 }
 
 function AlertTitle({
@@ -46,11 +46,11 @@ function AlertTitle({
       data-slot="alert-title"
       className={cn(
         'col-start-2 line-clamp-1 min-h-4 font-medium tracking-tight',
-        className
+        className,
       )}
       {...props}
     />
-  )
+  );
 }
 
 function AlertDescription({
@@ -62,11 +62,11 @@ function AlertDescription({
       data-slot="alert-description"
       className={cn(
         'text-muted-foreground col-start-2 grid justify-items-start gap-1 text-sm [&_p]:leading-relaxed',
-        className
+        className,
       )}
       {...props}
     />
-  )
+  );
 }
 
-export { Alert, AlertTitle, AlertDescription }
+export { Alert, AlertTitle, AlertDescription };
