+++ ayunis-core-frontend/src/widgets/model-distribution-chart/index.ts
@@ -1,0 +1,8 @@
+export { ModelDistributionWidget } from './ui/ModelDistributionWidget';
+export { buildModelDistributionChartData } from './lib/buildModelDistributionChartData';
+export type {
+  ChartDataItem,
+  ChartConfigEntry,
+  ModelBreakdownItem,
+  ModelDistributionChartData,
+} from './lib/buildModelDistributionChartData';
