+++ ayunis-core-frontend/src/widgets/model-distribution-chart/ui/ModelDistributionEmpty.tsx
