+++ ayunis-core-frontend/src/widgets/model-distribution-chart/ui/ModelDistributionError.tsx
