+++ ayunis-core-frontend/src/widgets/model-distribution-chart/ui/ModelDistributionLoading.tsx
