+++ ayunis-core-frontend/src/widgets/provider-consumption-chart/index.ts
@@ -1,0 +1,3 @@
+export { ProviderConsumptionWidget } from './ui/ProviderConsumptionWidget';
+export { buildProviderChartData } from './lib/buildProviderChartData';
+export type { ProviderChartData } from './lib/buildProviderChartData';
