+++ ayunis-core-frontend/src/widgets/provider-consumption-chart/ui/ProviderConsumptionChart.tsx
