+++ ayunis-core-frontend/src/widgets/provider-consumption-chart/ui/ProviderConsumptionError.tsx
