+++ ayunis-core-frontend/src/widgets/usage-stats-cards/index.ts
@@ -1,0 +1,1 @@
+export { UsageStatsCardsWidget } from './ui/UsageStatsCardsWidget';
