+++ ayunis-core-frontend/src/widgets/user-usage-table/index.ts
@@ -1,0 +1,4 @@
+export {
+  UserUsageTableWidget,
+  DEFAULT_PAGE_SIZE,
+} from './ui/UserUsageTableWidget';
